pprof server and acl check cache (#14)
* feat: add pprof server

* feat: add acl cache for mosquitto

* feat: remove task id from topic

+++ internal/bizapi/api/v1/log.go
@@ -6,7 +6,7 @@ import (
 
 type TaskLogReq struct {
 	g.Meta `path:"/task/log" method:"get"`
-	TaskId int `json:"taskId"  dc:"push task id, returned by successful /push response" v:"required#task id is required"`
+	TaskId int64 `json:"taskId"  dc:"push task id, returned by successful /push response" v:"required#task id is required"`
 }
 
 type TaskLogRes struct {
@@ -15,7 +15,7 @@ type TaskLogRes struct {
 
 type TaskStatusReq struct {
 	g.Meta `path:"/task/status" method:"get"`
-	TaskId int `json:"taskId"  dc:"push task id, returned by successful /push response" v:"required#task id is required"`
+	TaskId int64 `json:"taskId"  dc:"push task id, returned by successful /push response" v:"required#task id is required"`
 }
 
 type TaskStatusRes struct {
@@ -24,7 +24,7 @@ type TaskStatusRes struct {
 
 type PushLogReq struct {
 	g.Meta   `path:"/push/log" method:"get"`
-	TaskId   int    `json:"taskId" dc:"push task id" v:"required#push task id is required"`
+	TaskId   int64  `json:"taskId" dc:"push task id" v:"required#push task id is required"`
 	AppId    int    `json:"appId" dc:"app id" v:"required#app id is required"`
 	DeviceId string `json:"deviceId" dc:"device id" v:"required#device id is required"`
 }
